🚧♻️ Refactoring onboarding landing page: register.js
Dividing register.js into views and added more error catching for personal details and email
