+++ app/Domits/src/screens/home/screens/Home.js
@@ -67,7 +67,7 @@ const HomeScreen = () => {
 
         const response = await GetWishlist();
 
-        setFavorites(response.wishlists);
+        setFavorites(response.AccommodationIDs);
         setFavoritesLoading(false);
     }, []);
