+++ app/Domits/src/screens/login/hostonboardinglanding/styles/HostOnboardingLandingStyles.js
@@ -1,0 +1,124 @@
+import {StyleSheet} from "react-native";
+
+export const styles = StyleSheet.create({
+    highlightText: {
+        color: '#003366',
+    },
+    container: {
+        alignItems: 'center',
+    },
+    headerContainer: {
+        paddingTop: '20%',
+        alignItems: 'center',
+        paddingHorizontal: '5%',
+        paddingBottom: '3%',
+    },
+    header: {
+        fontSize: 35,
+        color: 'black',
+        fontWeight: 'bold',
+        textAlign: 'center',
+    },
+
+    singUpContainer:{
+        paddingTop: '5%',
+        color: 'black',
+        backgroundColor: 'white',
+        padding: 10,
+        borderRadius: 5,
+        width: '100%',
+        alignItems: 'center',
+        marginVertical: 5,
+
+    },
+
+    subHeader: {
+        fontSize: 25,
+        fontWeight: 'bold',
+        color: 'black',
+        marginTop: 4,
+        textAlign: 'center',
+    },
+    thirdHeader: {
+        fontSize: 30,
+        fontWeight: 'bold',
+        color: 'black',
+        marginTop: 4,
+        textAlign: 'center',
+    },
+    textBoldHostDomits: {
+        fontSize: 20,
+        fontWeight: 'bold',
+        color: 'black',
+        marginTop: 4,
+        textAlign: 'center',
+    },
+    Numbers: {
+        fontSize: 75,
+        color: '#003366',
+    },
+    description: {
+        fontWeight: 'bold',
+        fontSize: 15,
+        color: 'black',
+        marginTop: 8,
+        maxWidth: '90%',
+        paddingHorizontal: '5%',
+    },
+    button: {
+        backgroundColor: '#003366',
+        padding: 10,
+        marginVertical: 16,
+        borderRadius: 3,
+    },
+    buttonText: {
+        color: 'white',
+        fontSize: 18,
+    },
+    label: {
+        fontSize: 16,
+        color: 'black',
+        marginTop: 10,
+        alignSelf: 'flex-start',
+        paddingHorizontal: '5%',
+    },
+    input: {
+        color: 'black',
+        backgroundColor: 'white',
+        padding: 10,
+        borderRadius: 5,
+        width: '90%',
+        marginVertical: 5,
+        borderColor: '#003366',
+        borderWidth: 1,
+    },
+    checkboxContainer: {
+        flexDirection: 'row',
+        alignItems: 'center',
+        marginTop: 10,
+        width: '90%',
+    },
+    checkboxWrapper: {
+        width: '10%',
+    },
+    checkboxLabel: {
+        fontSize: 16,
+        color: 'black',
+        marginLeft: 8,
+        flexShrink: 1,
+    },
+    infoContainer: {
+        marginTop: 20,
+        alignItems: 'center',
+        backgroundColor: '#F9F9F9',
+        width: '100%',
+        paddingVertical: 20,
+    },
+    text: {
+        color: 'black',
+        fontSize: 16,
+        textAlign: 'center',
+        marginVertical: 2,
+        width: '75%',
+    },
+});
